Create SQLServer Unit Test
Took the existing MLOpsBuilderExtensionTests.cs that was stored in the integration tests and added a new SQLServer unit test project to align with the other storage tests.
